requirements.txt
streamlit
pandas
pyserial
opencv-python
numpy
Pillow
deepface
vosk
pyaudio
ollama
edge-tts
pygame